Added header-span definition support
The idea behind these spans is that the distance between each header represents
a set of GLAPI and #define calls. The idea is to only define header spans that
do not overlap onto other header spans. This makes each span represent a unique
part of the file. Currently, I do not have a way to detecting overlapping spans
but I am working on it.