Added the target-symbol-is-special function.
This function automates the process of defining symbol splitting due to the
fact that the body of each rule assigned to symbol splitting is the same except
for the target symbol. By adding this function I make it easier to maintain the
code as I add new features. It also doesn't matter if I construct a binary
image of the expert system because the rules will have already been defined.

The only downside is that batch* must be used instead of load*. While this may
seem like an issue, it really isn't.