In the process of fixing order dependency bugs in argument parsing
It turns out that since the matching of patterns goes "backwards" in how it
traverses a multifield the argument parsing is horribly broken.

To this end, it is necessary to rethink how to do this so that we can maintain
order without worrying about the order of the input even though the order of
the input is important.