Modified GLAPIConversion.clp to maintain an arg index in an unordered fashion
By using the number of preceeding commas + 1 it is possible to figure out the
position of a given argument without having to maintain state that would be
difficult to keep correct. The tradeoff is a linear count of the number of
commas for each argument on each call. This tradeoff only makes the code a
little less efficient while making it far easier to maintain.