Added parsing support for glu.h
I have already tested the code and each file is generates itself separately.
This is due to the "continuation" behavior that the MAIN module demonstrates
through the use of the focus command. The MAIN module is pushed onto the stack
first before the other modules. This allows any other parse facts to be
processed and the procedure to occur again.

This make the process of conversion very automated.