Deleted the id and class slots as well as the init message-handler
The slots consumed extra memory for no real benefit. In the case of id, the
name field already exists (it is an implicit field) and matches on
instance-name instead of symbol. This means that we store the instance-names in
a list instead of symbols. It also means that the name retrieved from an object
can be directly invoked without having to match against it (in the case where
that is all we were looking for).

The class slot is useless because we can just call (class ...) if it is
necessary. There is no need to waste space. Hell, for matching purposes we can
just use the (is-a ) field to do pattern matching as well.

The init message handler is no longer necessary because these fields no longer
exist.

I left the parent field in because I find it really useful in a wide number of
cases.