Deleted android.mk, module.mk, and many other components
I feel that android is not a proper platform to target for research because it
is so limited unless you use java which is nice if you actually want to use
Java. I like Java, but I feel that it would be a tad more effective if you
hollowed it out and filled it with beetles (at times at least).

The module.mk is for scummvm which is not a target for this platform (at least
at this time).

The Input folders are really unnecessary because getting input really isn't a
strength of an expert system such as CLIPS. Input processing is best left to
other languages.

The application, lib, and logic folders are a quaint idea but not really that
useful. The loadengine.clp file represents a tether that I can no longer stand.