Added basic support for loading programs in memory.
I am in the process of implementing operations that are actually objects in the
system. I am deciding whether the initial implementation should be
message-based (requires a subclass for each operation and a general pain in the
ass with the upside being that its easy to write "microcode" if I want to get
that low level) or using a common operation class that asserts facts which
contain the actual operation logic. I'm not sure which one to do at this point.
I'm leaning towards the more "pain-in-the-ass" method because its easier to
understand and I can rearrange encodings without having to change execution
rules. The only downside is that the "overhead" of message passing could come
into play. However, this is not a safety-critical or performance tool so who
cares....hmmmmmm

Anyway, I have also implemented a cacheless design for the processor due to the
way that the knowledge base works combined with the fact that it is dead
fucking simple to implement. Caches add a lot of complexity that really is not
necessary at this point.