Started defining the instruction set.
The instruction set is unique in that there is only one instruction that can
operate with immediates. Every other instruction operates exclusively on
registers. This has the side effect of writing programs in assembly a pain in
the ass. It also has a positive side effect of making it really easy to change
operations internally to speed them up.

For example,

set r134 1
add r212 r39 r134

would be compressed internally to

add r212 r39 1

or something similar when the program is running. This is legal because
this instruction is never found in user program and would have a very
different meaning. In fact, encoding {add} 212 39 1 would yield:

add r212 r39 r1

by the decoder.

However, at this point I just want to see if I can get the processor working
and executing programs before moving onto runtime optimization techniques.