Added the ability to assert facts into the expert system when invoked
These facts will be cleared once reset is called so I'm thinking of adding
deffacts assertion support to make it possible to reassert this information.
However, the problem is that how do you ensure that it's done correctly...

Ideally, args should be called after f and f2 to make sure that everything is
reset and loaded before continuing

Anyway, to assert command line arguments you go ./electron <file loads etc>
-args <args>

These args need to be surrounded by a string to make sure it is a single
element for the purposes of asserting it. It is also important to note that the
arguments are a string so that the user can decide what do with it. The fact
that is asserted is of the form (args <args>) where <args> is a string.