Fixed issues with the simulator relating to loads
It turns out that the merge function only works correctly if I wrap
multifields in a create$ operation. I will have to figure out what is
going on but for now the load operation is working correclty now.

I also added a new test case that tests out loading and storing